+++ TimberCommons/package/CHANGELOG.md
@@ -1,4 +1,5 @@
 # v1.6 (started on 2/7/2024):
+* [Fix #33] Water valve doesn't handle contamination correctly.
 
 # v1.5 (3 Feb 2024):
 * Relase irrigation towers code.